#51 gitlab gRPC logic moved to gitaly package.
#51 `/gitaly/server` package contains registration and start up logic for gitaly specific gRPC and provides start & stop interfaces.
